major: better reboot with heartwood

+++ support/reboot.sh
@@ -1,21 +1,11 @@
 #!/bin/bash
 
-shouldServeHeartwood=true
-
-# Parse arguments
-for arg in "$@"; do
-    case $arg in
-    --shouldServeHeartwood=*)
-        shouldServeHeartwood="${arg#*=}"
-        shift
-        ;;
-    *)
-        # If any other arguments are provided, call reboot-skill.sh
-        ./support/reboot-skill.sh "$@"
-        exit $?
-        ;;
-    esac
-done
+# Check if additional arguments are passed
+if [ "$#" -gt 0 ]; then
+    # Pass all arguments to reboot-skill.sh
+    ./support/reboot-skill.sh "$@"
+    exit $?
+fi
 
 yarn shutdown --shouldListRunning=false
 
@@ -25,13 +15,4 @@ hero "Resetting reboot counts..."
 
 ./support/pm2.sh reset all --silent
 
-# check if heartwood is installed at packages/spruce-heartwood-skill
-if [ ! -d packages/spruce-heartwood-skill ]; then
-    shouldServeHeartwood=false
-fi
-
-if [ "$shouldServeHeartwood" = true ]; then
-    yarn boot.serve
-else
-    yarn boot
-fi
+yarn boot
